pyproject.toml
[project]
name = "data-analyst-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.2.0",
    "ipython>=9.5.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.6",
    "matplotlib>=3.10.5",
    "mcp>=1.13.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
]
